Remove splash page
The fb login button is now on the main page
