+++ app/src/main/AndroidManifest.xml
@@ -11,18 +11,15 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme" >
-        <activity
-            android:name=".auth.SplashActivity"
-            android:label="@string/app_name" >
+
+        <activity android:name=".MainActivity"
+            android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
-        <activity android:name=".MainActivity"
-            android:label="@string/app_name" />
 
         <activity android:name="com.facebook.FacebookActivity"
             android:configChanges=
