+++ app/src/main/res/layout/activity_main.xml
@@ -19,4 +19,10 @@
         android:layout_marginTop="16dp"
         android:text="@string/create_post"
         android:textStyle="bold"/>
+
+    <fragment
+        android:id="@+id/login_fragment"
+        android:name="com.example.mona.facebookoffline.auth.LoginFragment"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content" />
 </LinearLayout>
